# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clivvy API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/shorts/{short_id}/audio": {
      "get": {
        "summary": "List Audio",
        "operationId": "list_audio_api_v1_shorts__short_id__audio_get",
        "parameters": [
          {
            "name": "short_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Short Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AudioAssetView"
                  },
                  "title": "Response List Audio Api V1 Shorts  Short Id  Audio Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload Audio",
        "operationId": "upload_audio_api_v1_shorts__short_id__audio_post",
        "parameters": [
          {
            "name": "short_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Short Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "music",
                "voice"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "filename",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "My audio",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioAssetView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shorts/{short_id}/audio/{asset_id}": {
      "get": {
        "summary": "Audio Media",
        "operationId": "audio_media_api_v1_shorts__short_id__audio__asset_id__get",
        "parameters": [
          {
            "name": "short_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Short Id"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/music/{track_id}/preview": {
      "get": {
        "summary": "Music Preview",
        "operationId": "music_preview_api_v1_music__track_id__preview_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AudioAssetView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "short_id": {
            "type": "string",
            "title": "Short Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "short_id",
          "kind",
          "filename",
          "duration_seconds"
        ],
        "title": "AudioAssetView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}